Fix size_t to int narrowing in TLS BIO bufcap
get_bio_bufcap() returns size_t but tls_bio_new() and bio_buf_init()
took int, so a bufcap above INT_MAX underwent an implementation-defined
conversion to a negative int that BUF_MEM_grow() then re-widened to
size_t, issuing an absurd allocation request. Values at or above 2^32
could also truncate to a small positive int and silently create
undersized buffers.

The capacity path now uses size_t end to end; an unallocatable bufcap
keeps surfacing as (nil, error) from accept()/connect().

+++ src/tls_bio.c
@@ -457,7 +457,7 @@ static int gc_lua(lua_State *L)
     return 0;
 }
 
-static inline int bio_buf_init(tls_bio_buf_t *b, int cap)
+static inline int bio_buf_init(tls_bio_buf_t *b, size_t cap)
 {
     b->mem = BUF_MEM_new();
     if (!b->mem) {
@@ -472,7 +472,7 @@ static inline int bio_buf_init(tls_bio_buf_t *b, int cap)
     return 0;
 }
 
-tls_bio_t *tls_bio_new(lua_State *L, int fd, int cap)
+tls_bio_t *tls_bio_new(lua_State *L, int fd, size_t cap)
 {
     tls_bio_t *bio = lua_newuserdata(L, sizeof(tls_bio_t));
